[distributed] moving to a new dRoby communication scheme
We drop the use of dRuby: we only use asynchronous communication and dRuby does
not help here (in fact, it does more harm than good), and is costly
performance-wise. Instead, we implement a TCP-based solution in which data is
marshalled by execution cycle (thus reducing the size of the marshalled data).
The cycles are processed directly in the control thread, and we have thus the
guarantee that all messages from a single execution cycle are processed at
once. This will allow to remove hacks for event propagation for instance.

For now, only simple message transmission works. TC_Communication mockup
infrastructure has been ported and test_transmit and test_call do work. There
is very few error handling (in particular, disconnection handling is broken)