handle wrong or incomplete index files
The first time the index file is loaded, the index data is checked
agaoinst the read log file by loading a 100 cycles that are registered
in the index files. If that check does not pass, the index file is
rebuilt.

This will make sure that a corrupted index file will be rebuilt.
Moreover, it will also rebuild an index file which is not the one of the
current log file (i.e. dowloaded a new event log with an index file
lying around)