log: remove a useless reference to Roby.engine
Roby.engine is nil until Roby::Application has been set up, so it is
most likely to be nil in Log::Server.enable_discovery. Moreover, the set
up of #disable_discovery in the finalization of Roby.engine is really a
task that should not be done by enable_discovery itself, but by the
setup code.