Added some more validation for emission calling of tasks.
Added the methods check_emission_validity and check_call_validity in event.rb
which are both called in call/emit and call/emit_without_propagation
for further error handling.

Extracted the exception refinement in task.rb for the
EventNotExecutable exception. The check_*_validity methods were overwritten
to make use of the refinement.

Added tests in test_task.rb to test the raising of the EventNotExecutable
exception.