doc: improving documentation
This commit adds the overview.rdoc file in the list of documents
generated by the doc:core target. Moreover, a big improvement has been
done on the documentation of the relation system. Finally, it is
mentioned in the Task class that the methods defined by the relations
are listed as methods of the TaskStructure module.