core: add a way to disable the "GC cannot be controlled" warning
This warning is useful on real controllers since the global
behaviour does depend on Roby's ability to control GC.
However, it is very annoying in tests.

An ExecutionEngine#gc_warning? flag is added by this commit,
allowing to disable it selectively. It is enabled by default,
and disabled in Roby::Test#setup