log: adapt error-reporting in FileLogger to the current logging method
Currently, logging is done per-cycle instead of per-message for
a huge gain in both performance and resulting size. The error
reporting in FileLogger#dump_method is not adapted to that and
always reports :cycle_end to be the faulty mesage.

This commit makes sure that it reports the actual message that
is not marshallable instead.