core: fix marshalling of PlanningTask and PlanningLoop
PlanningTask and PlanningLoop can now act on a "free method"
definition, i.e. a MethodDefinition that is not based on a
name but on a Proc object. These method definitions are not
marshallable, and therefore cannot be used in logging and
in multi-robot contexts.

We fix this commit fixes this by removing planning_method
from the set of arguments, and making sure that only the
meaningful arguments are marshalled. method_name is
reintroduced as a mandatory argument which is automatically
initialized from planning_method if needed. It is nil in
case of free methods (which is fine).