log: allow filtering of log messages before they are droby-marshalled
Before this commit, all messages where pushed to the log queue,
leaving the responsibility of filtering to loggers themselves.
This has the disadvantage that all events are droby-marshalled,
regardless of whether they will be used or not.

This commit fixes it by requesting a logs_message? method on
loggers. This method should return true for messages that
are to be logged and false otherwise. If no logger wants the
message, it is not pushed on the queue.