workaround a problem in exception handling in poll blocks
The problem is the following. Let's assume a user wants to both *emit* and
*react* to an event in the poll block. And he does it that way:

  poll do
    if success_condition
      emit :success
      ... react to success event ...
    end
  end

What would happen if the reaction has errors ? The generic poll implementation
will emit failed and we have a task that both failed and succeeded at the same
time.

The model would actually dictates that the reaction is done in an event handler.
But user's don't always follow the model.

So right now, I worked around a minimal side-effect of this problem, which is
MissionFailedError assuming that #terminal_event is the final event. Instead, we
actually use the history of +failed+.