rework documentation generation
This commit moves the rake code for the User's Guide generation
from doc/using/Rakefile into the main Rakefile, thus integrating
it with the main distribution system. Moreover, it adapts the
current Rake code from the standard RDoc to the rdoc gem (which
compatible with 1.9's rdoc system AFAIK).

The goal is in the end to move all the non-API documentation
into the webgen part (user's guide) and link to the API from there.