core: rework signalling/forwarding/event handlers interface
The following changes have been implemented:
 EventGenerator#on and Task#on can only be used to establish event
  handlers. The #signals methods have been added to establish signals.
 #signal has been renamed into #signals
 #forward has been renamed into #forward_to

Moreover, on tasks, the model-level .on method can also be used only to add
event handlers. For signals use .signal and for forwarding .forward

A warning is issued when the deprecated versions are called. The core
test suite has been updated, but not the dRoby test suite.

TEMP: signal/forward/... renaming