core: fix ChildFailedError in case of unreachability in tasks
Since 8f2e0e2790e9c4299372b720a479342e82480098, the hierarchy relation tests for
unreachability to know if a child has failed. The problem is that it makes the
failure point the event that became unreachable, instead of the event that
triggered the unreachability.

Fix this. This noticeably makes (for instance) the task's failure event the
failure point of a standard dependency relation.