properly handle planning threads that fail at transaction commit
This commit adds the proper code so that PlanningTask fails if
the call to #commit_transaction fails. It was previously finishing
with :success

That was happening because PlanningTask#poll was checking on freezed?
to know if #commit_transaction has been called, instead of actually checking
that the commit was successful. This commit adds the #committed? flag
to verify the latter.