doc: fix example usage of ThreadTask
Once one though it would be nice to send the result of the ThreadTask as
context of the :success event. Unfortunately, it would have a possibly
big performance impact when logging is on, so the idea was dropped.

Fix the "Basics" part of the guide which was using this inexistent
feature :P