gui: fix and use the #clear_integrated method
That method had been added before to better manage the cleanup of
old data in the plan displays. The thing is: it was disabled.

Resurrect it, which brings some simplifications and probably some
display bugfixes. Moreover, it fixes a bug introduced with
0a095de24704a69ed06cdb2371e6b9bf3e544a38. When an empty cycle was
encountered, the "integrated data" was not cleared. Thus, we did
not remove the signal propagation display after a "step" for
instance, which was confusing.