change how subrelations share the embedded information object
Until now, subrelations where sharing both the graph information
(edges) and the embedded information.

It seems to not be the best strategy. What is really important
is to share the edge info (obviously). But the embedded info
is semantically attached to the relation we're creating.

As a side-effect, it is now possible to do

class MyClass < Roby::Task
  def initialize(arguments)
    forward_to :start, self, :ready, :delay => 2
    super
  end
end

It was previously an error since task create a precedence relation
between start and all the other events.