properly handle a race condition in ExternalProcessTask
The problem is that the child can terminate before the
parent continues its execution. In this case, the SIGCHLD
handler will silently ignore the information, as no
ExternalProcessTask instance is yet associated with the
process.

The fix is to first register the instance with the process'
PID and *then* check that it is actually still running.