app: big improvement in the shell's usability
There is now a 'help' message and a way to describe the actions,
and display the description in the shell.

The planning methods are described by adding a "describe" statement
before the method's definition:

  describe("this method does this and that").
    required_arg("test", "if we are in test mode").
    optional_arg("blabla", "never set that")
  method("my_method") do
    ...
  end

Then, the information is available by using the shell's new
"actions", "actions_summary" and "help" commands.