app: map remote models locally in the shell
we create, on connection, local anonymous models for
the models of our remote controller. This is needed
for find_tasks to be useful.

This has the side effect to try to connect at startup
time. The shell will now display a user-understandable
message if the remote controller cannot be contacted and
exit with an error status.