core: fix a propagation ordering bug between forwards and signals
During propagation, it is more intuitive that the engine prefers forwards to
signals. Since forwards are meant as a way to generalize events while signals
are action propagation, we intuitively expect that -- for instance -- the
'stop' event of a task is emitted before the events that are signalled by
'success'.

So, prefer forwarding to signalling.

This patch may have a non-negligible performance effect unfortunately, so we
should watch out for performance problems.