core: be conservative for root tasks in Dependency#fullfilled_model
The old behaviour was to say that a root task was replaceable
by anything (i.e. the model it fullfilled was Roby::Task, {}.

That seems wrong: if there is a root task standing there (for instance,
a mission), then obviously it is there for a reason. The only sane
thing is therefore to return the most conservative answer
(self.model, self.arguments)