log: make the relation display aware of any kind of task relations
In particular, the display should not be limited to relations
defined in TaskStructure. To make loading other relations possible,
a new --require/-r options has been added to roby-log, making it
possible to load files before the display is started.

scripts/replay loads init.rb and therefore would not need to
do something like that.