roby: add roles and dependency merging to #depends_on
This commit adds a :roles key in the dependency information,
and the corresponding :role and :roles options to #depends_on
(the first one being simply a shortcut for :roles => [value]).
Roles are simply sets of strings for now.

Moreover, we add the ability to call #depends_on multiple times
between the same tasks, in which case the dependency constraints
are properly merged, or ModelViolation is raised if they are
inconsistent.