core: define #can_merge?
See documentation. The idea of a.can_merge?(b) is that it returns
true if +a+ can replace +b+, even though +a+ must be modified before
it replaces +b+. This is different to #can_replace?: can_replace?
returns true if +b+ can be replaced by +a+ with no modification of +a+