core: implement PlanObject#merged_relations
In transaction context, #merged_relations enumerates the related
objects of the receiver, in the context of the "applied" transaction.
I.e. it will enumerate the relations that the underlying plan would
have, if the transaction stack was completely applied. This is
different from #each_(child|parent)_object which only enumerates
the relations present in the transaction (local view vs. global
view)