core: don't proxy #fullfills? in task proxies
Proxy.proxy should not be used to convert containers other
than arrays, because it would change that container into an
array.

In practice, it meant that #fullfills was broken when arguments
were provided. However, #fullfills? should actually not be proxied
as we want the engine to take into account the arguments in the
transaction, not in the plan.