core: integrate the ability to merge relation information
We recently added the ability for the Dependency relation to
merge the relation information when it makes sense. That ability
was done in #depends_on directly, and therefore was not working
in case #replace_task was used.

This commit generically integrates this ability: if the developer
adds a .merge_info(from_object, to_object, old_info, new_info) method
to the relation object, then this method will be used to update the
relation information when it is applicable. If it returns nil, the
old behaviour of raising an exception is retained.