core: mark a task as failed? if emit_failed is called on :start
There is nothing that *forces* a task to call emit_failed on :start,
but if it does, the task is now marked as failed and should be respawned.

This is to avoid weird issues when a task is repeatedly started even
though the start fails. The engine should not do that by itself, but
the actual decision-making should be involved. Force the issue by
marking the task as failed.