core: improve support for copying objects around
Note: after this commit, copying objects becomes a potentially
dangerous operation (it was not really a nice one beforehand ...),
as the object is *not* automatically added back in the plan
anymore.

The rationale is that we are going to fix the transaction
proxying by copying objects and extending them with modules
instead of having the half-backed we have right now.