core: in MethodModel#model_of, use a concrete method if it is the only one under the name
Until now, either model_of was given with an 'id' argument, in which
case we were returning the model of the provided method, or we were
returning the global method model for the given method.

This also returns the concrete task model if there is only one
concrete task under the required name, i.e. for

  method :name, :returns => SimpleTask do
    ...
  end

"planner.model_of :name" will return SimpleTask while before
it was returning 'nil'