core: streamline exception handling from task code
This commit changes how code error in task code is handled,
making it more in-line with the overall error handling.

Previous commits changed the task-level handling of errors,
by having #emit_failed being always called whenever a
command/event handler fails. That did change how the task was
handled in the case where start! fails, but other than that
the error was still added in the engine's error handling and
caused all tasks that depend on it to fail.

This commit changes it by "simply" making the task fail whenever
a command/event handler fails. The error then gets propagated (and
displayed) only if the task was depended-upon in the plan, which is
much more in-line with the rest of error handling. The original
exception is made available in Task#failure_reason

In other words, the task also becomes a unit w.r.t. errors from the
handling of unexpected errors, isolating those errors from the rest
of the system.