app: add auto-saving of log/ into results/ and make the time tag on results/ more useful
scripts/run will now auto-save log/ into results/ (it can be disabled by setting
log.autosave to 'false'). Moreover, the time tag, that was only YYYYMMDD until
now has been changed to YYYYMMDD-HHMM *and* the actual time tag is the time at
which the controller started, not the time at which we save.