core:  fix duping task contexts and events
This commit splits the history of the original and the new
event, and makes sure that happened? is properly propagated.

Moreover, it applies the same policy to task events: until now,
fresh events were created, and now the events of the dup'ed task
are copies of the events of the original task.