core: in transactions, call #remove_object to remove the plan child objects
This fixes a bug in which an object that had been removed with #remove_object
could not be added again. The reason was that the plan child of the main object
(i.e. events for a task) were not removed from the transaction. Since we don't
commit them explicitely (the plan root has to), it was silently ignored.
However, if we start to add the removed task back into the transaction, the old
proxies were used thus leading to a "cannot add task back into the plan"