core: don't cache the return values of the mapping block in difference
Until now, we were computing a hash of all the values for mapping[]
for the specified tasks. However, the +tasks+ set given to #difference
is not complete, as we will also take into account these tasks neighbours
(which may not be in +tasks+ themselves).

Therefore, at least for now, do not cache the return values of the
given lambda.