revise ExternalProcess to make it more robust
First, we can't really detect start failures in #start_process,
as it is perfectly possible that the process finishes before
start_process returns.

Second, Ruby has a bug that makes the child of a fork() freeze.
We now use a timeout on the control channel to the child, so that
we can detect such situations and retry to start the process.