core: modify how ChildFailedError gets its failure_point
In case the explanation contains one event and zero or more generators,
we use the event as failure point.

The rationale is as follows. Even though the predicate is
explained by both the event emission *and* other event
non-emissions, we are actually interested in the emission
part (i.e. the error is that an event has been emitted).