core: fix how tasks are killed because of a relation error
What a relation error means is that a parent does not have its dependencies
met. In practice, it means that the parent can not continue to function. It
tells nothing about the child.

This therefore changes how tasks are killed because of unhandled relation
errors. Namely, only the parents (and the parent-of-parent) are killed, and the
child is left alone. It fixes issues with plans where a single task is used by
different subplans, each subplan having different requirements on the task.

Before, both subplans would have been terminated. Now, only the subplan that
has an unmet dependency is killed, the other one is left alone.