core: fix dependency constraints when the parent does not run and/or the child already runs
The dependency relation was broken in two situations:

First, if the child was already running. In that case, the relevant
events were not emitted and therefore the child was not checked
against the constraints in #check_structure.

Second, if the parent does not run. In that case, #check_structure
should not look at the relation at all (who cares if the parent does
not run), and we need to inform #check_structure that the child
should be checked when the parent starts.