tasks: don't use SIGCHLD for Tasks::ExternalProcess
Signal handling and multi-threading does not really goes together,
and Ruby 1.9 have some issues with it. Morevoer, trapping signals
in a framework leads to problems as other libraries may want to
do it as well (and we don't really need it).

Instead, add a propagation handler that calls Process.wait