core: transactions were duplicating event handlers
Any handler originally existing on an event was re-added to
the event after it has been wrapped in a transaction.

We use the same technique than with unreachable_handlers and
simply clear the @handlers set when we setup the proxy. New
handlers are then -- simply -- the ones that have been registered
afterwards

It has the drawback of "hiding" the current event handlers in
the transaction, but that is not a problem -- at least not for now.