core: fix handling of removals for single_child relations
The issue is that single child relations might, at some point at least,
have multiple children: in transactions.

Indeed, during the transaction commit, the object may have both a
transaction proxy and the corresponding plan object. The issue is then
that the caching implemented for single child relations was setting the
value to nil.

Instead, search all children in the relation to find out if we can set
the single child attribute to an existing child.