core: fix unreachable handlers when unreachable! is called from the propagation phase
This fixes the modification applied in 9077f3c9efe791ccbacbce4c9493f54e341ef433.

The commit was delaying calls to unreachability handlers until the next
propagation phase if the current phase was not propagation, but missed
out calling the handlers right away if the current phase is a propagation
phase.

This obviously broke quite a few things.