core: fix handling of so-called "additional errors"
The errors that are added during the exception propagation
itself are stored in the engine's additional_error set, and
then propagated themselves.

However, they could never end up in the fatal error set, which
is obviously wrong. Fix that.

Moreover, normalize the representation of fatal errors in the
execution loop, to avoid having a weird mix between a mapping
from exception objects to tasks and a simple list of exceptions.