core: *do* call poll blocks while the task is finishing
A recent change I made was to not call poll blocks if the task was finishing
(i.e. one terminal event was pending). This breaks most uses of #poll, as #poll
is often used to detect the end of the task (while failed! or stop! is used to
trigger its termination).

However, it seems that this change was meant to avoid calling #poll again while
it failed already once. Change the test on finished? to a test that checks if #poll
failed.