core: implement argument default values and delayed argument handling
The default values are given at declaration time

  argument :name, :default => value

if the value responds to #evaluate_delayed_argument, then the
argument's value is not set. Instead, the method is called to
evaluate the argument's value. The value is then set when the
task starts.

A DelayedTaskArgument class allows to do that through a block
(mainly for testing purposes though)

A side effect of this commit is that

  arguments :bla, :blo

is deprecated. The singular version

  argument :bla
  argument :blo

should be used instead