core: fix poll handler IDs deregistration
The handler ID for the task's poll blocks was shared across the
various instances of a single task model. In effect, it meant that
one of these tasks finishing was always deregistering the handler
of the last task started.

Fix this by storing the handler ID as an instance variable in the
task instances themselves.