core: extend the delayed arguments API
First, it is now possible to get arguments from the task's own
attributes. E.g., a planning task could get its goal argument
with

   from(:planned_task).goal_point

Second, the behaviour of the DelayedArgumentFromObject handler
when the specified argument is not available can be configured
through the 'weak' option. If true, it is simply ignored. If false,
it generates an error. This allows to create a from_state handler
(where it has to be "false" as the State object can be dynamically
extended)