core: separate the default dependent model in #depends_on from Task#model
This introduces the #provided_services call that is used by #depends_on
to get the default depended-upon model. It allows tasks to override
this default model.

In practice, it is designed to allow placeholders for task model tags,
i.e. task models that -- in themselves -- are only used to place a
task model tag in the plan. The task therefore only provides a model tag,
not the task model itself.