core: always display backtrace in Exception#pretty_print
I am going back and forth with displaying it always or never, but
can't really come up with a nice solution

The problem with always displaying it is that it is harder for caller
of #pretty_print to do special formatting for backtraces. The good side
is that one that "just" wants to pretty print an object of unknown
type (for instance, an event context), can do so without questions.