log: make the log format a bit more robust
We now prefix each marshal'ed block with the size of the block.
This makes receiving partial blocks (in client/server setups) a
LOT more robust as we don't have to try unmarshalling to know that
the block is partial (!)

This bumps the log file version to 4, and makes future changes to
the format easier to carry out - dumping and loading are concentrated
in two methods now.

Finally, the prologue is not written with Marshal anymore, but directly
read with normal methods, and contains a magic string.