core: change the relation addition API to integrate plan generation in a nice way
Until now, relation addition API always expected a task, e.g.

  parent.depends_on(child)

The API now accepts +child+ to either be a task, or an object that
responds to 'as_plan'. In the latter case, the #as_plan method is
called and should return the task that should be added.

This is meant to integrate plan generation APIs in a nicer way
(from the programmer's perspective). For instance,

  mission.depends_on(MoveTo)

will call MoveTo.as_plan, which can generate a semi-complex plan that
will generate path planning and select tasks.

Moreover, the #with_arguments call allows to specify arguments that
should be passed to the #as_plan method:

  mission.depends_on(MoveTo.with_arguments(:target => [10, 20]))

Beware that #with_arguments will raise NoMethodError if MoveTo does
not respond to :as_plan

Roby::Task defines an as_plan method that returns a new instance of
the task