log: fix overriding of task event models in the plan rebuilder
When an anonymous task model is created, it inherits its parent
task's event models. However, the remote model may have overriden
the controlable flag. This is something we do not know until we
get a message involving a task event of the overridden type.

To handle this, an event model overriding mechanism exists. However,
the following may happen:

 * task1 of model M is received, which leads to the creation of ev1
 * task2 of model M is received, which leads to the creation of ev1
 * ev1 of task1 is received, which leads to the stop event being
   updated on task1 and M
 * ev1 of task2 is received

The issue is that updating task1 and M do not update ev1's model
on task2. So, before this commit, the last point was leading the
code to try and override ev1's model on M again, which is forbidden.

Instead, we only have to update the model of ev1.