core: change the structure of the complete execution phase
This commit results of the realization that the current execution
phase breaks the synchronicity assumption a bit too much.

Indeed, we now have the following parts of the system that are
involved during the execution cycle:

 * the event propagation
 * the scheduler
 * the garbage collector
 * the exception handling (through the garbage collector)

However, due to the structure of the system, some events that
could be scheduled after the event propagation are not, and
might then be garbage collected. The same kind of who-is-first-wins
situations goes for the garbage collection vs. next event propagation
phases and so on.

It also leads to slow reaction times if multiple levels of tasks
are needed and the scheduler is heavily used (as the scheduler is
called only once per cycle).

Overall, this leads to a pretty fragile execution, as the result
might start to heavily depend on the execution times of each
individual processes, which is something we DEFINITELY want to
avoid in the model (recall: we want "pseudo-synchronous" execution)

So, the execution cycle is changed to process ALL events that are
internally generated, regardless of their source, while keeping the
propagation / error handling / GC separation.

How this works in practice is that initial events are gathered ONCE
from the outside of the system (external processes, time-related
triggers). They are then propagated through in an inner loop where
scheduling and GCing is integrated:

  gather_external_events
  while there_is_something_to_propagate
     propagate_all_events
     if implicit_event_triggering # this is where scheduling is integrated
         next
     else
         do_error_handling
         do_garbage_collection
     end
  end