core: allow to override the abstract flag on a per-instance basis
Abstract is used to mark "placeholders", i.e. tasks that are present
in the plan as a mark for future actions that need to be planned/
developped.

However, one sometimes uses non-abstract models to represent these
placeholders. Currently, this is done by setting executable to false
on these tasks. But the representation is then ambiguous: is executable?
false because of the current system state or because the task is a
placeholder ?

Therefore, allow to override abstract? on a per-instance basis to remove
this ambiguity.