core: support for copying event handlers during replacements
This commit adds options to the EventGenerator#on method, to allow
marking event handlers as "copy-on-replace", i.e. as being copied
to a new task if the current task gets replaced.

The idea behind that is that one can add event handlers to a placeholder
("abstract") task in the plan, and that these handlers get propagated
to the task that is being used later on. This is actually done transparently:
the copy-on-replace flag is set to true by default on tasks that are
abstract.