core: fix instance-level poll block support in transactions, and implement :on_replace => :copy
Poll blocks could not be added in transactions (as event handlers can),
so fix that. Moreover, add the same :on_replace option than for event
handlers, that allow to control how poll blocks get copied when a task
gets replaced.

This feature is turned on for abstract tasks by default, as it allows
to add poll blocks to abstract tasks and then have the poll block
actually executed on the real task.