log: fix initial connection to the process server
The code was calling read_prologue on the event_file IO object,
but that object is also used to track the amount of data read so
far. This was obviously a bit of an issue.

So, we now properly call read_prologue on the new_data string
(i.e. data read) through a StringIO.