log: do not do unnecessary deep copies when replaying the logs
The strategy now is that a given task is shared across all the plans
that contain it. It is the duty of the display to check what is the
task's state at a given point in time.

This has two advantages: we don't need to deep copy the whole plan
each time, but only the plan lists themselves. This improves speed,
display responsiveness and memory usage tremendously.

Moreover, it will allow other types of displays as e.g. chronicles
to be implemented quite easily