core: fix deadlock if the log thread dies
The communication between the core execution engine and the log thread
is done through a SizedQueue. This queue has the property that, when it
gets full, the core execution engine will pass to the log thread to empty
it.

In addition, when the log thread dies, it announces that logging is impossible
by setting @logging_thread to nil (which makes logging? return false).

However, Log.log was *not* looking for logging?, which led it to continue
pushing data to the queue and, in the end, deadlocking