log: properly handle garbage-collected tasks
The purpose of this commit is to make sure that GCed tasks are
displayed *with relations* in the plan display. Until now, these
tasks were displayed, but without something marking that they are
begin finalized and -- more importantly -- with the relations already
removed. This was really a problem for people to understand the plans.