core: make sure that a planningfailederror leads to the removal of the planned task
A recent change made PlanningFailedError point to the planning task,
not the planned task (and rightly so). The issue is that, then, the
planned task was not getting removed.

Explicitely tell the exception handling system that the
planningfailederror, while originating from the planning task, should
be propagated from the planned task.