core: fix ordering of task status predicate update w.r.t. event emission
The status updates were done after the event handlers got called,
which caused the very unintuitive

  on :stop do
    finished? => false
  end

This fixes the issue and adds a more precise test to check for this
kind of issues.

The problem was added in a previous commit that reordered the
calls so that unreachability gets updated after the event is
registered in the task history. The test covers this as well.