core: define Task#execute
This is meant as a tool for code that modified running systems.

The issue is the single-time execution of a block if a task is
running. On "pending plan" modifications, one would use "on :start".

However, on tasks that might already be running, this does not work
as the start event got already emitted. #execute covers both cases in
one call, avoiding mistakes.