core: fix execution agent respawning
There was a conflict between a recent (generally wanted) change
and an old (sometimes wanted) feature.

The recent change is to add a DecisionControl default behaviour
on execution-agent-failed-for-pending-task: to mark the task as
failed to start.

The old feature is that, when a model-level agent is added, new
agents would be restarted automatically.

This patch reconciles both by making the respawning feature optional.
The default behaviour is now to have an error, but if the model-level
agent is added with :respawn => true, the respawning behaviour is
added instead.