core: fix error handling relation
The error handling is now called more than once per cycle, and
within event propagation context. We must therefore handle the
repairing tasks that are either in running or starting states,
as well as call start! right away instead of through once { }

The issue there was that the event was not propagated properly
in tests. This is a corner case, but could have happened on real
systems as well:

  add repair task
  start it in once { }

--> once { } gets called only next cycle. If another error handling
    pass is triggered, two things could happen: either the same task
    is added as error handler and multiple once blocks will be executed
    (Harmless but not very nice). More problematic could be that another
    of the available repair tasks gets added to the plan and boom.