core: integrate error handling in synchronous event processing
This integrates the error handling pass in the event processing
mechanisms used by #call and #emit when no event loop is running.

This is critical to allow the usage of exception handling as a
"normal" error translation mechanism within the framework itself
(more on that in later commits). Moreover, it makes the tests more
strict as calling an event that causes a structure error will now
raise an exception (and therefore cannot be overlooked)