log: fix plan replay w.r.t. Plan#clear being a lot more strict
Plan#clear now calls the finalization codepath for every task and
event included in the plan, which leads to sanity checking on these
plan objects. However, the trick in plan replay is to always keep
the same task object across all plans in which it is included. So,
plans end up including tasks that have already been finalized and
Plan#clear is not happy about it.

Overload #finalize_object to do the strict minimum (i.e. clear the
relations).