core: fix interaction between scheduling and task start failures
The emission failure of task's start events does not instantaneously mark the
task as failed to start anymore. This led to having the scheduler try to
schedule the task over and over again without the error being caught.

Workaround by not attempting to schedule an event for which an error is
queued. In general, we should probably either ignore all propagations
for such events, or raise an exception.