core: fix relationship btw. pending? and failed calls
The issue here is that, so far, a failed call would remove the pending
flag. However, until the error is propagated, this is harmful (we don't
know how the error is going to be managed.

So, keep the pending flag to true until

This fixes an interaction issue between scheduling and error handling
in commands: a task whose command has failed would be eligible for
scheduling, which would loop forever.