state: create a state model structure parallel to the one from our superclass
The reason is that we do

  state.pose.position = value

If we return for state.pose the value of state.superclass.pose, we
are going to add position on the parent state model structure instead
of doing it on the child. So, create a new StateFieldModel for state.pose,
making sure that it has the "right" superclass (i.e. superclass.state.pose)