state: allow modifying the model after the StateSpace instance got initialized
This is required so that State works as it always worked (i.e. backward
compatibility). One must now do State.new_model on it to add a model
to it (and make it a lot stricter in the process)

This finally cleans up some of the behind-the-scene initialization code
by making it a lot more straightforward and less "magic"