actions: define the action interface, a very stripped down version of the planners
Most of the functionality present in the planners is actually unused,
at the price of hard to understand code and fragility.

This action interface will replace the planners in the future. It is (1) better
architectured and (2) a lot simpler.

The planner code and app integration is kept, to keep backward compatibility.
New apps will generate action interfaces instead of planners, however.