actions: support passing arguments to the state machines
The arguments are provided in the state machine definition as
"state variables", i.e. placeholders that look like local variables
but are replaced at instanciation time by their values.

Additionally, objects that respond to either #to_action_state
or to #instanciate can be used as states, but need to be converted
explicitly with #make_state