reduce some pathological performance problems when adding objects to big plans
There are two issues.

One is that we copy sets whose size is of the same order of
magnitude than known_tasks, and we do that per addition. This
ends up costing quite a lot for big plans.

Moreover, we do set difference operations that are also O(M+N).
Since we know that the addition set is usually pretty small,
these operations are replaced by O(M.log(N))

  small_set.delete_if { big_set.include?(t) }

A quick check on gnuplot shows that if the set of new tasks / new
events is smaller or equal than 8, the MlogN operation wins always.
Otherwise, up to ~15, it is at worst 1.25 times slower if the
big set is somewhere in between 10 and 30 elements and better
otherwise.

The operation we really optimize for is when adding a single task,
so the MlogN operation is really a winner there.