core: migrate task scripting to Actions::ExecutionContextModel
We gain some load-time checking of the scripts (for task models that
respond to find_child and find_event), but lose the ability to late-bind
on childs and events in #wait if one of the children in the chain
is abstract and being planned. This case seemed very "dark magicky"
to me, and I am therefore not-so unhappy to remove it.