core: modify behaviour of #can_merge?
can_merge? was checking the fullfilled model as the target model
that should be fullfill-able by self. However, #merge
is meant to work on the actually selected tasks (not what they are
selected for), so #can_merge? could return true while #merge would
raise.

Change #can_merge? to match the semantics of #merge