coordination: the child model resolved from the model task must take precedence w.r.t. the instance child
Scripts are really meant as objects that work on the model more than
on the instance. As such, if the task models provide child information,
we must use this information first and

In the context of syskit, the situation is the one of a composition
child that, in the model, is a data service and in reality is a
real component. From the point of view of the script, we want the
task to behave as the data service instead of behaving as the actual
task (w.r.t. port mappings for instance)